+++ src/clauses/Match.test.ts
@@ -22,13 +22,14 @@ import Cypher from "..";
 
 describe("CypherBuilder Match", () => {
     test("Match node", () => {
-        const movieNode = new Cypher.Node({
-            labels: ["Movie"],
-        });
+        const movieNode = new Cypher.Node();
 
         const matchQuery = new Cypher.Match(
-            new Cypher.Pattern(movieNode).withProperties({
-                test: new Cypher.Param("test-value"),
+            new Cypher.Pattern(movieNode, {
+                labels: ["Movie"],
+                properties: {
+                    test: new Cypher.Param("test-value"),
+                },
             })
         );
 
@@ -43,13 +44,11 @@ describe("CypherBuilder Match", () => {
     });
 
     test("Match node with return passing a clause", () => {
-        const movieNode = new Cypher.Node({
-            labels: ["Movie"],
-        });
+        const movieNode = new Cypher.Node();
 
         const returnClause = new Cypher.Return(movieNode);
 
-        const matchQuery = new Cypher.Match(movieNode).return(returnClause);
+        const matchQuery = new Cypher.Match(new Cypher.Pattern(movieNode, { labels: ["Movie"] })).return(returnClause);
 
         const queryResult = matchQuery.build();
         expect(queryResult.cypher).toMatchInlineSnapshot(`
@@ -64,11 +63,9 @@ describe("CypherBuilder Match", () => {
         const idParam = new Cypher.Param("my-id");
         const nameParam = new Cypher.Param("my-name");
 
-        const movieNode = new Cypher.Node({
-            labels: ["Movie"],
-        });
+        const movieNode = new Cypher.Node();
 
-        const matchQuery = new Cypher.Match(movieNode)
+        const matchQuery = new Cypher.Match(new Cypher.Pattern(movieNode, { labels: ["Movie"] }))
             .where(movieNode, { id: idParam, name: nameParam })
             .remove(movieNode.property("name"))
             .return(movieNode.property("id"));
@@ -93,11 +90,9 @@ describe("CypherBuilder Match", () => {
         const idParam = new Cypher.Param("my-id");
         const nameParam = new Cypher.Param("my-name");
 
-        const movieNode = new Cypher.Node({
-            labels: ["Movie"],
-        });
+        const movieNode = new Cypher.Node();
 
-        const matchQuery = new Cypher.Match(movieNode)
+        const matchQuery = new Cypher.Match(new Cypher.Pattern(movieNode, { labels: ["Movie"] }))
             .where(movieNode, { id: idParam, name: nameParam })
             .remove(movieNode.property("name"), movieNode.property("released"))
             .return(movieNode.property("id"));
@@ -119,12 +114,13 @@ describe("CypherBuilder Match", () => {
     });
 
     test("Optional match", () => {
-        const movieNode = new Cypher.Node({
-            labels: ["Movie"],
-        });
+        const movieNode = new Cypher.Node();
 
-        const pattern = new Cypher.Pattern(movieNode).withProperties({
-            test: new Cypher.Param("test-value"),
+        const pattern = new Cypher.Pattern(movieNode, {
+            labels: ["Movie"],
+            properties: {
+                test: new Cypher.Param("test-value"),
+            },
         });
         const matchQuery = new Cypher.OptionalMatch(pattern).return(movieNode);
 
@@ -177,11 +173,13 @@ RETURN this1"
     describe("Assign to path variable", () => {
         const a = new Cypher.Node();
         const b = new Cypher.Node();
-        const rel = new Cypher.Relationship({
-            type: "ACTED_IN",
-        });
+        const rel = new Cypher.Relationship();
 
-        const pattern = new Cypher.Pattern(a).related(rel).to(b);
+        const pattern = new Cypher.Pattern(a)
+            .related(rel, {
+                type: "ACTED_IN",
+            })
+            .to(b);
 
         test("with unique id", () => {
             const path = new Cypher.Path();
@@ -216,13 +214,14 @@ RETURN this1"
             const nameParam = new Cypher.Param("my-name");
             const ageParam = new Cypher.Param(5);
 
-            const movieNode = new Cypher.Node({
-                labels: ["Movie"],
-            });
+            const movieNode = new Cypher.Node();
 
             const matchQuery = new Cypher.Match(
-                new Cypher.Pattern(movieNode).withProperties({
-                    test: new Cypher.Param("test-value"),
+                new Cypher.Pattern(movieNode, {
+                    labels: ["Movie"],
+                    properties: {
+                        test: new Cypher.Param("test-value"),
+                    },
                 })
             )
                 .where(movieNode, { id: idParam, name: nameParam, age: ageParam })
@@ -250,13 +249,14 @@ RETURN this1"
             const nameParam = new Cypher.Param("my-name");
             const ageParam = new Cypher.Param(5);
 
-            const movieNode = new Cypher.Node({
-                labels: ["Movie"],
-            });
+            const movieNode = new Cypher.Node();
 
             const matchQuery = new Cypher.Match(
-                new Cypher.Pattern(movieNode).withProperties({
-                    test: new Cypher.Param("test-value"),
+                new Cypher.Pattern(movieNode, {
+                    labels: ["Movie"],
+                    properties: {
+                        test: new Cypher.Param("test-value"),
+                    },
                 })
             )
                 .where(movieNode, { id: idParam, name: nameParam, age: ageParam })
@@ -284,13 +284,14 @@ RETURN this1"
             const idParam = new Cypher.Param("my-id");
             const nameParam = new Cypher.Param("my-name");
 
-            const movieNode = new Cypher.Node({
-                labels: ["Movie"],
-            });
+            const movieNode = new Cypher.Node();
 
             const matchQuery = new Cypher.Match(
-                new Cypher.Pattern(movieNode).withProperties({
-                    test: new Cypher.Param("test-value"),
+                new Cypher.Pattern(movieNode, {
+                    labels: ["Movie"],
+                    properties: {
+                        test: new Cypher.Param("test-value"),
+                    },
                 })
             )
                 .where(movieNode, { id: idParam, name: nameParam })
@@ -315,13 +316,14 @@ RETURN this1"
         test("Match with null values", () => {
             const testParam = new Cypher.Param(null);
 
-            const movieNode = new Cypher.Node({
-                labels: ["Movie"],
-            });
+            const movieNode = new Cypher.Node();
 
             const matchQuery = new Cypher.Match(
-                new Cypher.Pattern(movieNode).withProperties({
-                    test: testParam,
+                new Cypher.Pattern(movieNode, {
+                    labels: ["Movie"],
+                    properties: {
+                        test: testParam,
+                    },
                 })
             )
                 .where(Cypher.isNull(movieNode.property("name")))
@@ -338,10 +340,10 @@ RETURN this1"
         });
 
         test("Match Where with complex operation", () => {
-            const node = new Cypher.Node({ labels: ["Movie"] });
+            const node = new Cypher.Node();
 
             const param = new Cypher.Param(1);
-            const clause = new Cypher.Match(node)
+            const clause = new Cypher.Match(new Cypher.Pattern(node, { labels: ["Movie"] }))
                 .where(
                     Cypher.and(
                         Cypher.or(Cypher.gt(param, new Cypher.Param(2)), Cypher.lt(param, new Cypher.Param(4))),
@@ -369,9 +371,9 @@ RETURN this1"
         });
 
         test("Match where with property and not", () => {
-            const node = new Cypher.Node({ labels: ["Movie"] });
+            const node = new Cypher.Node();
 
-            const queryMatch = new Cypher.Match(node).where(
+            const queryMatch = new Cypher.Match(new Cypher.Pattern(node, { labels: ["Movie"] })).where(
                 Cypher.not(Cypher.eq(node.property("title"), new Cypher.Param("Matrix")))
             );
 
@@ -391,11 +393,13 @@ RETURN this1"
         test("Match node with simple NOT", () => {
             const nameParam = new Cypher.Param("my-name");
 
-            const movieNode = new Cypher.Node({
-                labels: ["Movie"],
-            });
+            const movieNode = new Cypher.Node();
 
-            const matchQuery = new Cypher.Match(movieNode)
+            const matchQuery = new Cypher.Match(
+                new Cypher.Pattern(movieNode, {
+                    labels: ["Movie"],
+                })
+            )
                 .where(Cypher.not(Cypher.eq(movieNode.property("name"), nameParam)))
                 .return(movieNode);
 
@@ -417,11 +421,9 @@ RETURN this1"
             const nameParam = new Cypher.Param("my-name");
             const ageParam = new Cypher.Param(5);
 
-            const movieNode = new Cypher.Node({
-                labels: ["Movie"],
-            });
+            const movieNode = new Cypher.Node();
 
-            const matchQuery = new Cypher.Match(movieNode)
+            const matchQuery = new Cypher.Match(new Cypher.Pattern(movieNode, { labels: ["Movie"] }))
                 .where(
                     Cypher.not(
                         Cypher.or(
@@ -448,11 +450,11 @@ RETURN this1"
         });
 
         test("Match where with empty operation", () => {
-            const movieNode = new Cypher.Node({
-                labels: ["Movie"],
-            });
+            const movieNode = new Cypher.Node();
 
-            const matchQuery = new Cypher.Match(movieNode).where(new Cypher.Raw("")).return(movieNode);
+            const matchQuery = new Cypher.Match(new Cypher.Pattern(movieNode, { labels: ["Movie"] }))
+                .where(new Cypher.Raw(""))
+                .return(movieNode);
 
             const queryResult = matchQuery.build();
             expect(queryResult.cypher).toMatchInlineSnapshot(`
@@ -464,13 +466,13 @@ RETURN this1"
         });
 
         test("Match where with undefined predicate", () => {
-            const movieNode = new Cypher.Node({
-                labels: ["Movie"],
-            });
+            const movieNode = new Cypher.Node();
 
             const maybePredicate: Predicate | undefined = undefined;
 
-            const matchQuery = new Cypher.Match(movieNode).where(maybePredicate).return(movieNode);
+            const matchQuery = new Cypher.Match(new Cypher.Pattern(movieNode, { labels: ["Movie"] }))
+                .where(maybePredicate)
+                .return(movieNode);
 
             const queryResult = matchQuery.build();
             expect(queryResult.cypher).toMatchInlineSnapshot(`
@@ -481,13 +483,11 @@ RETURN this0"
             expect(queryResult.params).toMatchInlineSnapshot(`{}`);
         });
         test("Match where ... and with undefined predicate", () => {
-            const movieNode = new Cypher.Node({
-                labels: ["Movie"],
-            });
+            const movieNode = new Cypher.Node();
 
             const maybePredicate: Predicate | undefined = undefined;
 
-            const matchQuery = new Cypher.Match(movieNode)
+            const matchQuery = new Cypher.Match(new Cypher.Pattern(movieNode, { labels: ["Movie"] }))
                 .where(Cypher.eq(new Cypher.Variable(), new Cypher.Variable()))
                 .and(maybePredicate)
                 .return(movieNode);
@@ -508,11 +508,9 @@ RETURN this0"
             const nameParam = new Cypher.Param("Keanu Reeves");
             const evilKeanu = new Cypher.Param("Seveer unaeK");
 
-            const personNode = new Cypher.Node({
-                labels: ["Person"],
-            });
+            const personNode = new Cypher.Node();
 
-            const matchQuery = new Cypher.Match(personNode)
+            const matchQuery = new Cypher.Match(new Cypher.Pattern(personNode, { labels: ["Person"] }))
                 .where(personNode, { name: nameParam })
                 .set([personNode.property("name"), evilKeanu])
                 .return(personNode);
@@ -540,11 +538,9 @@ RETURN this0"
             const idParam = new Cypher.Param("my-id");
             const nameParam = new Cypher.Param("my-name");
 
-            const movieNode = new Cypher.Node({
-                labels: ["Movie"],
-            });
+            const movieNode = new Cypher.Node();
 
-            const matchQuery = new Cypher.Match(movieNode)
+            const matchQuery = new Cypher.Match(new Cypher.Pattern(movieNode, { labels: ["Movie"] }))
                 .where(movieNode, { id: idParam, name: nameParam })
                 .delete(movieNode)
                 .return(new Cypher.Literal(5));
@@ -569,11 +565,9 @@ RETURN this0"
             const idParam = new Cypher.Param("my-id");
             const nameParam = new Cypher.Param("my-name");
 
-            const movieNode = new Cypher.Node({
-                labels: ["Movie"],
-            });
+            const movieNode = new Cypher.Node();
 
-            const matchQuery = new Cypher.Match(movieNode)
+            const matchQuery = new Cypher.Match(new Cypher.Pattern(movieNode, { labels: ["Movie"] }))
                 .where(movieNode, { id: idParam, name: nameParam })
                 .detachDelete(movieNode);
 
@@ -596,11 +590,9 @@ RETURN this0"
             const idParam = new Cypher.Param("my-id");
             const nameParam = new Cypher.Param("my-name");
 
-            const movieNode = new Cypher.Node({
-                labels: ["Movie"],
-            });
+            const movieNode = new Cypher.Node();
 
-            const matchQuery = new Cypher.Match(movieNode)
+            const matchQuery = new Cypher.Match(new Cypher.Pattern(movieNode, { labels: ["Movie"] }))
                 .where(movieNode, { id: idParam, name: nameParam })
                 .noDetachDelete(movieNode);
 
@@ -622,17 +614,13 @@ RETURN this0"
 
     describe("Nested Match", () => {
         test("Match.match()", () => {
-            const movie1 = new Cypher.Node({
-                labels: ["Movie"],
-            });
+            const movie1 = new Cypher.Node();
 
-            const movie2 = new Cypher.Node({
-                labels: ["Movie"],
-            });
+            const movie2 = new Cypher.Node();
 
-            const matchQuery = new Cypher.Match(movie1)
+            const matchQuery = new Cypher.Match(new Cypher.Pattern(movie1, { labels: ["Movie"] }))
                 .where(Cypher.eq(movie1.property("title"), new Cypher.Param("movie1")))
-                .match(new Cypher.Pattern(movie2))
+                .match(new Cypher.Pattern(movie2, { labels: ["Movie"] }))
                 .where(Cypher.eq(movie1.property("title"), new Cypher.Param("movie2")));
 
             const queryResult = matchQuery.build();
@@ -652,19 +640,15 @@ WHERE this0.title = $param1"
         });
 
         test("Match.match() passing an existing Match clause", () => {
-            const movie1 = new Cypher.Node({
-                labels: ["Movie"],
-            });
+            const movie1 = new Cypher.Node();
 
-            const movie2 = new Cypher.Node({
-                labels: ["Movie"],
-            });
+            const movie2 = new Cypher.Node();
 
-            const secondMatch = new Cypher.Match(movie2).where(
+            const secondMatch = new Cypher.Match(new Cypher.Pattern(movie2, { labels: ["Movie"] })).where(
                 Cypher.eq(movie1.property("title"), new Cypher.Param("movie2"))
             );
 
-            const matchQuery = new Cypher.Match(movie1)
+            const matchQuery = new Cypher.Match(new Cypher.Pattern(movie1, { labels: ["Movie"] }))
                 .where(Cypher.eq(movie1.property("title"), new Cypher.Param("movie1")))
                 .match(secondMatch);
 
@@ -685,17 +669,13 @@ WHERE this0.title = $param1"
         });
 
         test("Match.optionalMatch()", () => {
-            const movie1 = new Cypher.Node({
-                labels: ["Movie"],
-            });
+            const movie1 = new Cypher.Node();
 
-            const movie2 = new Cypher.Node({
-                labels: ["Movie"],
-            });
+            const movie2 = new Cypher.Node();
 
-            const matchQuery = new Cypher.Match(movie1)
+            const matchQuery = new Cypher.Match(new Cypher.Pattern(movie1, { labels: ["Movie"] }))
                 .where(Cypher.eq(movie1.property("title"), new Cypher.Param("movie1")))
-                .optionalMatch(new Cypher.Pattern(movie2))
+                .optionalMatch(new Cypher.Pattern(movie2, { labels: ["Movie"] }))
                 .where(Cypher.eq(movie1.property("title"), new Cypher.Param("movie2")));
 
             const queryResult = matchQuery.build();
@@ -717,13 +697,11 @@ WHERE this0.title = $param1"
 
     describe("Match.callProcedure", () => {
         test("Match.match()", () => {
-            const movie1 = new Cypher.Node({
-                labels: ["Movie"],
-            });
+            const movie1 = new Cypher.Node();
 
             const labels = new Cypher.Variable();
 
-            const matchQuery = new Cypher.Match(movie1)
+            const matchQuery = new Cypher.Match(new Cypher.Pattern(movie1, { labels: ["Movie"] }))
                 .where(Cypher.eq(movie1.property("title"), new Cypher.Param("movie1")))
                 .callProcedure(new Cypher.Procedure("db.labels"))
                 .yield(["label", labels])
@@ -742,6 +720,29 @@ RETURN var1"
   "param0": "movie1",
 }
 `);
+        });
+    });
+
+    describe("Match.call", () => {
+        test("Match.Call with variable scope", () => {
+            const movieNode = new Cypher.Node();
+            const actorNode = new Cypher.Node();
+
+            const match = new Cypher.Match(new Cypher.Pattern(movieNode, { labels: ["Movie"] }))
+                .match(new Cypher.Pattern(actorNode, { labels: ["Actor"] }))
+                .call(new Cypher.Create(new Cypher.Pattern(movieNode).related().to(actorNode)), [movieNode, actorNode])
+                .return(movieNode);
+
+            const { cypher, params } = match.build();
+            expect(cypher).toMatchInlineSnapshot(`
+"MATCH (this0:Movie)
+MATCH (this1:Actor)
+CALL (this0, this1) {
+    CREATE (this0)-[this2]->(this1)
+}
+RETURN this0"
+`);
+            expect(params).toMatchInlineSnapshot(`{}`);
         });
     });
 });