+++ src/expressions/functions/string.ts
@@ -24,7 +24,18 @@ import { normalizeExpr } from "../../utils/normalize-variable";
 import { CypherFunction } from "./CypherFunctions";
 
 /**
- * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/)
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/#functions-btrim)
+ * @group Cypher Functions
+ * @category String
+ */
+export function btrim(original: string | Expr, trimCharacter?: string | Expr): CypherFunction {
+    const normalizedOriginal = normalizeExpr(original);
+    const normalizedTrimCharacter = trimCharacter ? normalizeExpr(trimCharacter) : undefined;
+    return new CypherFunction("btrim", filterTruthy([normalizedOriginal, normalizedTrimCharacter]));
+}
+
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/#functions-left)
  * @group Cypher Functions
  * @category String
  */
@@ -42,12 +53,26 @@ export function lower(original: Expr): CypherFunction {
 }
 
 /**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/#functions-ltrim)
+ * @group Cypher Functions
+ * @category String
+ * @deprecated Use {@link ltrim} instead
+ */
+export function lTrim(original: Expr | string, trimCharacter?: string | Expr): CypherFunction {
+    const normalizedOriginal = normalizeExpr(original);
+    const normalizedTrimCharacter = trimCharacter ? normalizeExpr(trimCharacter) : undefined;
+    return new CypherFunction("lTrim", filterTruthy([normalizedOriginal, normalizedTrimCharacter]));
+}
+
+/**
  * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/)
  * @group Cypher Functions
  * @category String
  */
-export function lTrim(original: Expr): CypherFunction {
-    return new CypherFunction("lTrim", [original]);
+export function ltrim(original: Expr | string, trimCharacter?: string | Expr): CypherFunction {
+    const normalizedOriginal = normalizeExpr(original);
+    const normalizedTrimCharacter = trimCharacter ? normalizeExpr(trimCharacter) : undefined;
+    return new CypherFunction("ltrim", filterTruthy([normalizedOriginal, normalizedTrimCharacter]));
 }
 
 /**
@@ -85,9 +110,23 @@ export function right(original: Expr, length: Expr): CypherFunction {
  * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/)
  * @group Cypher Functions
  * @category String
+ * @deprecated Use {@link rtrim} instead
  */
-export function rTrim(original: Expr): CypherFunction {
-    return new CypherFunction("rTrim", [original]);
+export function rTrim(original: Expr | string, trimCharacter?: string | Expr): CypherFunction {
+    const normalizedOriginal = normalizeExpr(original);
+    const normalizedTrimCharacter = trimCharacter ? normalizeExpr(trimCharacter) : undefined;
+    return new CypherFunction("rTrim", filterTruthy([normalizedOriginal, normalizedTrimCharacter]));
+}
+
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/)
+ * @group Cypher Functions
+ * @category String
+ */
+export function rtrim(original: Expr | string, trimCharacter?: string | Expr): CypherFunction {
+    const normalizedOriginal = normalizeExpr(original);
+    const normalizedTrimCharacter = trimCharacter ? normalizeExpr(trimCharacter) : undefined;
+    return new CypherFunction("rtrim", filterTruthy([normalizedOriginal, normalizedTrimCharacter]));
 }
 
 /**