+++ tests/deprecated/cdc.test.ts
@@ -17,9 +17,9 @@
  * limitations under the License.
  */
 
-import Cypher from "../../src/index";
+import Cypher from "../../src";
 
-describe("CDC procedures", () => {
+describe("CDC procedures - Deprecated", () => {
     test("cdc.current", () => {
         const query = Cypher.cdc.current();
         const { cypher } = query.build();